chore: stricter typing + broadened lint (P1 #10)
- Enable mypy check_untyped_defs and fix the 16 deprecated Qt enum accesses it
  surfaces (Qt.AlignCenter -> Qt.AlignmentFlag.AlignCenter, QEvent.Leave ->
  QEvent.Type.Leave, QMessageBox.Critical -> QMessageBox.Icon.Critical, etc.).
- Broaden ruff: add SIM + RUF (ignoring noisy RUF001-003 unicode + SIM105). Fixes
  incl. a real dead-code bug (p = p if p.suffix else p — a no-op), collection
  unpacking, redundant int() casts, ClassVar on constant lookup tables, and a
  collapsible-if. Removed a stray misplaced comment in media.py.

89 tests pass; ruff + mypy(check_untyped_defs) clean on 16 files.

+++ blender_discover.py
@@ -65,7 +65,7 @@ def discover_settings() -> dict:
         r = scene.render
         # Effective fps (accounts for NTSC bases like 23.976 → fps 24 / base 1.001).
         base = getattr(r, "fps_base", 1.0) or 1.0
-        s["fps"] = int(round(r.fps / base))
+        s["fps"] = round(r.fps / base)
         s["frame_start"] = int(scene.frame_start)
         s["frame_end"] = int(scene.frame_end)
         s["frame_step"] = int(getattr(scene, "frame_step", 1))
